requirements.txt
Python==3.11.7
pandas==2.0.0
matplotlib==3.7.1
numpy==1.24.2
scikit-learn==1.2.2
scikit-elm==0.21a0
requests==2.31.0
anvil-uplink==0.4.2
seaborn==0.12.2
os==3.11.7
lightgbm==4.2.0
prophet==1.1.5
holidays==0.41
catboost==1.2.2
